+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 > [!NOTE]
 > While we strive to maintain backwards compatibility as much as possible, we can't guarantee semantic versioning will be strictly followed, as this provider depends on the underlying [bunny.net API](https://docs.bunny.net/reference/bunnynet-api-overview).
 
+## [Unreleased]
+### Fixed
+- resource pullzone_hostname: create: clean up after a loadFreeCertificate error ([#36](https://github.com/BunnyWay/terraform-provider-bunnynet/issues/36));
+
 ## [0.7.0] - 2025-04-15
 ### Added
 - Support for Bunny Shield